Create README.md
A keylogger captures keystrokes from the keybaord once its is activated in the system. This application is designed to help cyber officers to track any kind of malicious activities that an attacker might carry out through his network. The programming language used in this project is Python and its various packages have been imported to execute the code.